packages/intervo-backend/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.27
langchain-openai>=0.0.8
langchain-community>=0.0.24
langchain-text-splitters>=0.0.1
langchain-voyageai>=0.0.3
langchain-google-genai>=0.0.7
langchain-groq>=0.1.2
langchain-deepseek>=0.1.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.1
typing-extensions>=4.9.0
pydantic>=2.6.3
aiohttp>=3.9.1
numpy>=1.26.4
tqdm>=4.66.2
fastapi>=0.110.0
uvicorn>=0.27.1
boto3>=1.34.51
botocore>=1.34.51
python-multipart>=0.0.7
python-json-logger>=2.0.7
httpx>=0.27.0
beautifulsoup4>=4.12.2
ratelimit>=2.2.1
lxml>=4.9.3
html5lib>=1.1
mammoth>=1.6.0
pdfplumber>=0.10.3 